smart-milk/analysis service/requirements.txt
paho-mqtt
mysql-connector-python
pandas